+++ hutool-captcha/src/test/java/cn/hutool/captcha/CaptchaTest.java
@@ -29,7 +29,7 @@ public class CaptchaTest {
 		// 定义图形验证码的长和宽
 		LineCaptcha lineCaptcha = CaptchaUtil.createLineCaptcha(200, 100);
 		lineCaptcha.setGenerator(new MathGenerator());
-		lineCaptcha.write("e:/code.png");
+		lineCaptcha.write("e:/captcha/math.png");
 	}
 
 	@Test
@@ -40,13 +40,13 @@ public class CaptchaTest {
 		LineCaptcha lineCaptcha = CaptchaUtil.createLineCaptcha(200, 100);
 		// LineCaptcha lineCaptcha = new LineCaptcha(200, 100, 4, 150);
 		// 图形验证码写出，可以写出到文件，也可以写出到流
-		lineCaptcha.write("e:/line.png");
+		lineCaptcha.write("e:/captcha/line.png");
 		Console.log(lineCaptcha.getCode());
 		// 验证图形验证码的有效性，返回boolean值
 		lineCaptcha.verify("1234");
 
 		lineCaptcha.createCode();
-		lineCaptcha.write("e:/line.png");
+		lineCaptcha.write("e:/captcha/line.png");
 		Console.log(lineCaptcha.getCode());
 		// 验证图形验证码的有效性，返回boolean值
 		lineCaptcha.verify("1234");
@@ -60,7 +60,7 @@ public class CaptchaTest {
 		CircleCaptcha captcha = CaptchaUtil.createCircleCaptcha(200, 100, 4, 20);
 		// CircleCaptcha captcha = new CircleCaptcha(200, 100, 4, 20);
 		// 图形验证码写出，可以写出到文件，也可以写出到流
-		captcha.write("e:/circle.png");
+		captcha.write("e:/captcha/circle.png");
 		// 验证图形验证码的有效性，返回boolean值
 		captcha.verify("1234");
 	}
@@ -73,7 +73,7 @@ public class CaptchaTest {
 		ShearCaptcha captcha = CaptchaUtil.createShearCaptcha(200, 100, 4, 4);
 		// ShearCaptcha captcha = new ShearCaptcha(200, 100, 4, 4);
 		// 图形验证码写出，可以写出到文件，也可以写出到流
-		captcha.write("e:/shear.png");
+		captcha.write("e:/captcha/shear.png");
 		// 验证图形验证码的有效性，返回boolean值
 		captcha.verify("1234");
 	}